requirements.txt
Django==5.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
django-modeltranslation==0.19.15
django-allauth==0.63.0
django-axes==6.1.0
django-otp==1.5.0
django-redis==5.4.0
django-storages==1.14
django-celery-beat==2.7.0
django-simple-history==3.10.1
django-filter==24.3
django-ratelimit==4.1.0
drf-spectacular==0.27.0
mysqlclient==2.2.4
celery==5.4.0
redis==5.0.0
requests==2.32.2
boto3==1.35.0
sentry-sdk>=2.32.0
Pillow==10.4
pytest-cov==5.0
python-decouple==3.8
pytest==8.3.0
pytest-django==4.9.0
pre-commit==3.8
black==24.8.0
flake8==7.1.0
argon2-cffi==23.1.0
geocoder==1.38.0
locust==2.31
isort==5.13
